+++ extensions/leetcode/src/problem-search.tsx
@@ -4,6 +4,8 @@ import { useMemo, useState } from 'react';
 import { endpoint, getProblemQuery, searchProblemQuery } from './api';
 import { GetProblemResponse, Problem, ProblemDifficulty, ProblemPreview, SearchProblemResponse } from './types';
 import { formatProblemMarkdown } from './utils';
+import { useCodeSnippets } from './hooks/useCodeSnippets';
+import { useProblemTemplateActions } from './hooks/useProblemTemplateActions';
 
 function formatDifficultyColor(difficulty: ProblemDifficulty): Color {
   switch (difficulty) {
@@ -19,7 +21,7 @@ function formatDifficultyColor(difficulty: ProblemDifficulty): Color {
 }
 
 function ProblemDetail(props: { titleSlug: string }): JSX.Element {
-  const { isLoading, data: problem } = useFetch<GetProblemResponse, undefined, Problem>(endpoint, {
+  const { isLoading: isProblemLoading, data: problem } = useFetch<GetProblemResponse, undefined, Problem>(endpoint, {
     method: 'POST',
     body: JSON.stringify({
       query: getProblemQuery,
@@ -37,29 +39,23 @@ function ProblemDetail(props: { titleSlug: string }): JSX.Element {
     },
   });
 
+  const { isLoading: isSnippetsLoading, data: codeSnippets } = useCodeSnippets(props.titleSlug);
+
   const problemMarkdown = useMemo(() => formatProblemMarkdown(problem), [problem]);
+
+  const actions = useProblemTemplateActions({
+    codeSnippets,
+    problemMarkdown,
+    isPaidOnly: problem?.isPaidOnly,
+    linkUrl: `https://leetcode.com/problems/${props.titleSlug}`
+  });
 
   return (
     <Detail
-      isLoading={isLoading}
+      isLoading={isProblemLoading || isSnippetsLoading}
       markdown={problemMarkdown}
-      actions={
-        <ActionPanel>
-          <Action.OpenInBrowser title="Open in Browser" url={`https://leetcode.com/problems/${props.titleSlug}`} />
-          <Action.CopyToClipboard
-            title="Copy Link to Clipboard"
-            content={`https://leetcode.com/problems/${props.titleSlug}`}
-          />
-          {!problem?.isPaidOnly && (
-            <Action.CopyToClipboard
-              title="Copy Problem to Clipboard"
-              content={problemMarkdown}
-              shortcut={{ modifiers: ['cmd'], key: 'c' }}
-            />
-          )}
-        </ActionPanel>
-      }
-    ></Detail>
+      actions={actions}
+    />
   );
 }